move avatar one level lower because we'll need it in the session picker
