+++ src/domain/avatar.js
